Always translate colors, not only for known props
Split the validation code in two: One to convert generally invalid stuff
to tmThemes such as functions and hash strings with the wrong number of
characters; and one to check if the (converted) type is valid for the
property. Color name replacement falls to the latter since you may use
idents that happen to have a color name in arbitrary properties.

Fixes #1.